refactor: 🛠  code refactoring and moving to files
